Event loop and drone acceleration calibration

+++ src/Timer.h
@@ -11,5 +11,5 @@ class Timer {
         Timer();
 
         void start();
-        float stop();
+        long long stop();
 };
